FFI: Split jet and test bindings
Bindings required to run jets remain in ffi.rs and are always compiled.
Bindings required to run tests are moved to tests/ffi.rs. The tests
module is gated behind the already existing test-utils feature.
Update fuzz crate to use test FFI.
