+++ fuzz/Cargo.toml
@@ -11,7 +11,7 @@ cargo-fuzz = true
 
 [dependencies]
 honggfuzz = { version = "0.5.55", default-features = false }
-simplicity-lang = { path = ".." }
+simplicity-lang = { path = "..", features = ["test-utils"] }
 
 [[bin]]
 name = "c_rust_merkle"
