+++ simplicity-sys/src/lib.rs
@@ -3,4 +3,5 @@
 pub mod c_jets;
 pub use c_jets::{CElementsTxEnv, CFrameItem, CTapEnv, CTransaction};
 pub mod ffi;
+#[cfg(feature = "test-utils")]
 pub mod tests;
